+++ stories/cart/CartQuantityHandler.stories.tsx
@@ -4,14 +4,14 @@ import { useState } from 'react'
 
 // More on default export: https://storybook.js.org/docs/react/writing-stories/introduction#default-export
 export default {
-  title: 'Example/CartQuantityHandler',
+  title: 'cart/CartQuantityHandler',
   component: CartQuantityHandler,
   // More on argTypes: https://storybook.js.org/docs/react/api/argtypes
   argTypes: {},
 } as ComponentMeta<typeof CartQuantityHandler>
 
 // More on component templates: https://storybook.js.org/docs/react/writing-stories/introduction#using-args
-const Template: ComponentStory<typeof CartQuantityHandler> = (args) => {
+const Template: ComponentStory<typeof CartQuantityHandler> = () => {
   const [quantity, setQuantity] = useState<number>(1)
   return (
     <CartQuantityHandler
